+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.8.1
 - Added Hide Blacklisted Recipe to Favorited Recipes
+- Change Craft Stack's tooltip to default state on
 - Fixed some features not been Unloaded in the right conditions
 - Fixed labels in the Reporting Bugs config